+++ lib/lightning_web/live/user_confirmation_required_live.ex
@@ -70,7 +70,7 @@ defmodule LightningWeb.UserConfirmationRequiredLive do
   def handle_event("validate_email", %{"user" => user_params}, socket) do
     changeset =
       socket.assigns.current_user
-      |> Accounts.validate_change_user_email(user_params)
+      |> Accounts.validate_change_user_email(user_params, validate_password: false)
       |> Map.put(:action, :validate_email)
 
     {:noreply, assign(socket, :email_changeset, changeset)}
@@ -119,4 +119,11 @@ defmodule LightningWeb.UserConfirmationRequiredLive do
          )}
     end
   end
+
+  def email_form_submit_disabled?(changeset) do
+    password = Ecto.Changeset.get_field(changeset, :current_password)
+
+    not changeset.valid? or password in [nil, ""]
+  end
+
 end